+++ CHANGELOG.md
@@ -38,6 +38,10 @@
 
 #### Builtin
 
+- **Breaking Change**: rename some ``LineEditor`` methods
+    - ``setPrompt`` -> ``setPrompter``
+    - ``setCompletion`` -> ``setCompleter``
+    - ``setHistSync`` -> ``setAcceptor``
 - overhaul internal keycode handling of ``LineEditor``
     - correctly read unrecognized escape sequences (CSI/SS3)
     - ``LineEditor#bind`` method accept human-readable keyname (such as `F1`, `ctrl+alt+a`,